Split graph into discrete component
